Move ChunkInfo to chunks file
