+++ hdf5/src/hl.rs
@@ -1,5 +1,4 @@
 pub mod attribute;
-#[cfg(feature = "1.13.0")]
 pub mod chunks;
 pub mod container;
 pub mod dataset;
