+++ hdf5/src/lib.rs
@@ -77,10 +77,10 @@ mod export {
     }
 
     pub mod dataset {
+        #[cfg(feature = "1.10.5")]
+        pub use crate::hl::chunks::ChunkInfo;
         #[cfg(feature = "1.13.0")]
         pub use crate::hl::chunks::ChunkInfoBorrowed;
-        #[cfg(feature = "1.10.5")]
-        pub use crate::hl::dataset::ChunkInfo;
         pub use crate::hl::dataset::{Chunk, Dataset, DatasetBuilder};
         pub use crate::hl::plist::dataset_access::*;
         pub use crate::hl::plist::dataset_create::*;
